Add authorization to application

+++ backend/routes/posts.js
@@ -38,6 +38,7 @@ router.post(
       title: req.body.title,
       content: req.body.content,
       imagePath: url + '/images/' + req.file.filename,
+      creator: req.userData.userId,
     });
     post.save().then((createdPost) => {
       res.status(201).json({
@@ -63,19 +64,31 @@ router.put(
       title: req.body.title,
       content: req.body.content,
       imagePath: imagePath,
+      creator: req.userData.userId,
     });
-    Post.updateOne({ _id: req.params.id }, post).then(() => {
-      res
-        .status(200)
-        .json({ message: 'Post updated successfully!', imagePath: imagePath });
+    Post.updateOne(
+      { _id: req.params.id, creator: req.userData.userId },
+      post
+    ).then((result) => {
+      if (result.n > 0) {
+        res.status(200).json({ message: 'Post updated successfully!' });
+      } else {
+        res.status(401).json({ message: 'Not authorized!' });
+      }
     });
   }
 );
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(() => {
-    res.status(200).json({ message: 'Post deleted!' });
-  });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(
+    (result) => {
+      if (result.n > 0) {
+        res.status(200).json({ message: 'Post deleted successfully!' });
+      } else {
+        res.status(401).json({ message: 'Not authorized!' });
+      }
+    }
+  );
 });
 
 router.get('', (req, res, next) => {
